test 10

+++ app/Http/Action/Product/ProductAction.php
@@ -6,6 +6,7 @@ use App\Http\Repository\Product\Read\ProductReadRepositoryInterface;
 use App\Http\Repository\Product\Write\ProductWriteRepositoryInterface;
 use Illuminate\Database\Eloquent\Collection as EloquentCollection;
 use Illuminate\Support\Collection;
+use Illuminate\Support\Facades\Gate;
 
 class ProductAction
 {
@@ -23,6 +24,10 @@ class ProductAction
 
     public function storeData(Collection $data): EloquentCollection
     {
+        if (auth()->user()->cannot('create', Product::class)) {
+            abort(403);
+        }
+        //Gate::authorize('create');
         return $this->productWriteRepository->Save($data);
     }
 
@@ -33,11 +38,13 @@ class ProductAction
 
     public function updateData(Collection $data, int $id): EloquentCollection
     {
+        Gate::authorize('update', '');
         return $this->productWriteRepository->update($data, $id);
     }
 
     public function deleteData(int $id)
     {
+        Gate::authorize('delete');
         return $this->productWriteRepository->delete($id);
     }
 }
